requirements.txt
langchain
langgraph>=0.0.32
langchain_community
langchain-ollama
ollama
python-dotenv
httpx 
pydantic

